Add in mypy and fix type hints
