+++ .gitignore
@@ -32,3 +32,7 @@ build/
 
 # SQLite
 *.db
+
+# caches
+.ruff_cache/
+.mypy_cache/
