+++ requirements-dev.txt
@@ -38,6 +38,10 @@ multidict==6.0.5
     #   -r requirements.txt
     #   aiohttp
     #   yarl
+mypy==1.11.0
+    # via -r requirements-dev.in
+mypy-extensions==1.0.0
+    # via mypy
 nodeenv==1.9.1
     # via pre-commit
 platformdirs==4.2.2
@@ -54,6 +58,7 @@ typing-extensions==4.12.2
     # via
     #   -r requirements.txt
     #   aiosqlite
+    #   mypy
 uv==0.2.25
     # via -r requirements-dev.in
 virtualenv==20.26.3
